Create Dokka plugin in desktop module.
