+++ desktop/build.gradle.kts
@@ -24,25 +24,10 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import io.gitlab.arturbosch.detekt.Detekt
-import org.jetbrains.dokka.DokkaConfiguration
-import org.jetbrains.dokka.base.DokkaBase
-import org.jetbrains.dokka.base.DokkaBaseConfiguration
-import org.jetbrains.dokka.gradle.DokkaTask
-import java.time.LocalDate
-
-buildscript {
-    dependencies {
-        // Required to customize HTML documents generated by Dokka.
-        classpath("org.jetbrains.dokka:dokka-base:1.9.20")
-    }
-}
-
 plugins {
     kotlin("jvm").version("1.9.20")
     id("org.jetbrains.compose").version("1.6.11")
-    id("io.gitlab.arturbosch.detekt").version("1.23.6")
-    id("org.jetbrains.dokka").version("1.9.20")
+    id("io.gitlab.arturbosch.detekt")
 }
 
 /**
@@ -90,68 +75,4 @@ dependencies {
     implementation("androidx.compose.runtime:runtime:1.6.0")
     implementation(compose.desktop.currentOs)
     implementation("io.spine.examples.pingh:client:$pinghVersion")
-}
-
-/**
- * After applying, Detekt is configured to use
- * `${parentRootDir}/config/quality/detekt-config.yml` file.
- */
-detekt {
-    buildUponDefaultConfig = true
-    config.from(files("$parentRootDir/config/quality/detekt-config.yml"))
-}
-
-/**
- * Specifies that Detekt generates only HTML reports.
- */
-tasks.withType<Detekt>().configureEach {
-    reports {
-        html.required.set(true)
-        xml.required.set(false)
-        txt.required.set(false)
-        sarif.required.set(false)
-        md.required.set(false)
-    }
-}
-
-/**
- * Configures [DokkaTask]s to accept Kotlin files and customizes output files.
- */
-tasks.withType<DokkaTask>().configureEach {
-    dokkaSourceSets.configureEach {
-        /**
-         * Configures links to the external Java documentation.
-         */
-        jdkVersion.set(jvmVersion)
-
-        skipEmptyPackages.set(true)
-
-        documentedVisibilities.set(
-            setOf(
-                DokkaConfiguration.Visibility.PUBLIC,
-                DokkaConfiguration.Visibility.PROTECTED
-            )
-        )
-
-        /**
-         * Dokka Base plugin allows to set a few properties to customize the output:
-         *
-         * - `separateInheritedMembers` when set to `true`, creates a separate tab in
-         * type-documentation for inherited members;
-         * - `footerMessage` property to specify footer text in page generated by Dokka.
-         *
-         * @see <a href="https://kotlinlang.org/docs/dokka-html.html#customize-assets">
-         *  Dokka modifying frontend assets</a>
-         */
-        pluginConfiguration<DokkaBase, DokkaBaseConfiguration> {
-            separateInheritedMembers = true
-            footerMessage = "Copyright ${LocalDate.now().year}, TeamDev"
-        }
-    }
-
-    /**
-     * Specifies output directory for documents generated by the Dokka.
-     */
-    val buildDirectory = layout.buildDirectory.asFile.get().path
-    outputDirectory.set(file("$buildDirectory/docs/dokkaKotlin"))
 }
